all: abstract fixtures, convert StateTest into a BlockchainTest generator
