+++ src/ethereum_test_tools/common/__init__.py
@@ -24,25 +24,18 @@ from .helpers import (
     to_hash,
     to_hash_bytes,
 )
+from .json import to_json
 from .types import (
     AccessList,
     Account,
     Address,
     Alloc,
     Auto,
-    Block,
     Bloom,
     Bytes,
     Environment,
-    Fixture,
-    FixtureBlock,
-    FixtureEngineNewPayload,
-    FixtureHeader,
     Hash,
-    Header,
     HeaderNonce,
-    HiveFixture,
-    InvalidFixtureBlock,
     JSONEncoder,
     Number,
     Removable,
@@ -53,7 +46,6 @@ from .types import (
     alloc_to_accounts,
     serialize_transactions,
     str_or_none,
-    to_json,
     withdrawals_root,
 )
 
@@ -65,22 +57,14 @@ __all__ = (
     "AddrBB",
     "Alloc",
     "Auto",
-    "Block",
     "Bloom",
     "Bytes",
     "EngineAPIError",
     "EmptyTrieRoot",
     "Environment",
-    "Fixture",
-    "FixtureBlock",
-    "FixtureEngineNewPayload",
-    "FixtureHeader",
     "Hash",
-    "Header",
     "HeaderNonce",
     "HistoryStorageAddress",
-    "HiveFixture",
-    "InvalidFixtureBlock",
     "JSONEncoder",
     "Number",
     "Removable",
