--- src/ethereum_test_tools/filling/__init__.py
@@ -1,6 +1,0 @@
-"""
-Test filling methods.
-"""
-from .fill import fill_test
-
-__all__ = ("fill_test",)
