+++ src/ethereum_test_tools/spec/__init__.py
@@ -1,11 +1,12 @@
 """
 Test spec definitions and utilities.
 """
-from .base_test import BaseTest, BaseTestConfig, TestSpec, verify_post_alloc
-from .blockchain_test import BlockchainTest, BlockchainTestFiller, BlockchainTestSpec
-from .state_test import StateTest, StateTestFiller, StateTestSpec
+from .base.base_test import BaseFixture, BaseTest, BaseTestConfig, TestSpec, verify_post_alloc
+from .blockchain.blockchain_test import BlockchainTest, BlockchainTestFiller, BlockchainTestSpec
+from .state.state_test import StateTest, StateTestFiller, StateTestSpec
 
 __all__ = (
+    "BaseFixture",
     "BaseTest",
     "BaseTestConfig",
     "BlockchainTest",
