+++ src/ethereum_test_tools/spec/base/__init__.py
@@ -1,0 +1,3 @@
+"""
+Base test spec class and utilities
+"""
