+++ src/ethereum_test_tools/tests/test_code.py
@@ -13,7 +13,6 @@ from evm_transition_tool import GethTransitionTool
 
 from ..code import CalldataCase, Case, Code, Conditional, Initcode, Switch, Yul
 from ..common import Account, Environment, TestAddress, Transaction, to_hash_bytes
-from ..filling import fill_test
 from ..spec import StateTest
 from ..vm.opcode import Opcodes as Op
 from .conftest import SOLC_PADDING_VERSION
@@ -648,9 +647,9 @@ def test_switch(tx_data: bytes, switch_bytecode: bytes, expected_storage: Mappin
     txs = [Transaction(to=code_address, data=tx_data, gas_limit=1_000_000)]
     post = {TestAddress: Account(nonce=1), code_address: Account(storage=expected_storage)}
     state_test = StateTest(env=Environment(), pre=pre, txs=txs, post=post)
-    fill_test(
+    state_test.base_test_config.blockchain_test = True
+    state_test.generate(
         t8n=GethTransitionTool(),
-        test_spec=state_test,
         fork=Shanghai,
-        spec=None,
+        eips=None,
     )
